[Jingtao] change the type of result to long to avoid overflow

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/FibonacciCalculator.java
@@ -1,13 +1,13 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class FibonacciCalculator {
-    public static int calculate(int position){
+    public static long calculate(int position){
         if(position <= 2){
             return 1;
         }
-        int nextResult;
-        int currentResult = 1;
-        int lastResult = 1;
+        long nextResult;
+        long currentResult = 1;
+        long lastResult = 1;
         for (int i=2;i<position;i++){
             nextResult = currentResult + lastResult;
             lastResult = currentResult;
